change file structure and renaming
