--- python-lib/ocr_constants.py
@@ -1,15 +1,0 @@
-class Constants:
-    OCR_TYPES = ["pdf", "jpg", "jpeg", "png", "tiff"]
-    PDF_MULTI_SUFFIX = "_pdf_page_"
-    FUNCTIONS_DEF = "functions_definition"
-    PIPELINE_DEF = "pipeline_definition"
-    DPI = "dpi"
-    QUALITY = "quality"
-    RECOMBINE_PDF = "recombine_pdf"
-    LANGUAGE_TESSERACT = "language"
-    LANGUAGE_EASYOCR = "language_easyocr"
-    OCR_ENGINE = "ocr_engine"
-    DEFAULT_ENGINE = "default"
-    TESSERACT = "tesseract"
-    EASYOCR = "easyocr"
-    EASYOCR_READER = "easyocr_reader"
