+++ python-lib/text_extraction_ocr_utils/__init__.py
@@ -1,15 +1,28 @@
-import docx
 from io import BytesIO
 import logging
-from ocr_constants import Constants
 import os
-import pypandoc
 import pypdfium2 as pdfium
 from shutil import which
-import tempfile
 
 
 logger = logging.getLogger(__name__)
+
+
+class Constants:
+    OCR_TYPES = ["pdf", "jpg", "jpeg", "png", "tiff"]
+    PDF_MULTI_SUFFIX = "_pdf_page_"
+    FUNCTIONS_DEF = "functions_definition"
+    PIPELINE_DEF = "pipeline_definition"
+    DPI = "dpi"
+    QUALITY = "quality"
+    RECOMBINE_PDF = "recombine_pdf"
+    LANGUAGE_TESSERACT = "language"
+    LANGUAGE_EASYOCR = "language_easyocr"
+    OCR_ENGINE = "ocr_engine"
+    DEFAULT_ENGINE = "default"
+    TESSERACT = "tesseract"
+    EASYOCR = "easyocr"
+    EASYOCR_READER = "easyocr_reader"
 
 
 def pdf_to_pil_images_iterator(pdf_bytes, dpi=None):
@@ -51,7 +64,7 @@ def image_processing_parameters(recipe_config):
     return params
 
 
-def ocr_text_extraction_parameters(recipe_config):
+def ocr_parameters(recipe_config):
     """ retrieve text extraction recipe parameters """
     params = {}
     params[Constants.RECOMBINE_PDF] = recipe_config.get(Constants.RECOMBINE_PDF, False)
@@ -87,47 +100,3 @@ def get_default_ocr_engine():
     if which("tesseract") is not None:  # check if tesseract is in the path
         return Constants.TESSERACT
     return Constants.EASYOCR
-
-
-def try_download_pandoc():
-    """ download pandoc prebuilt binaries into the temporary job folder """
-    try:
-        import pypandoc
-        pandoc_tmp_directory = os.getcwd()
-        pypandoc.download_pandoc(targetfolder=pandoc_tmp_directory, download_folder=pandoc_tmp_directory)
-        return True
-    except Exception as e:
-        logger.warning("Failed to download pandoc binaries: {}".format(e))
-        return False
-
-
-def extract_text(file_bytes, extension, with_pandoc):
-    """
-    First extract with pdfium or docx PDFs and docx files.
-    Then try to extract text using pandoc into plain text.
-    Finally, just decode the bytes if pandoc failed or is not downloaded.
-    """
-    if extension == "pdf":
-        pdf_pages = pdfium.PdfDocument(file_bytes)
-        return "\n".join([page.get_textpage().get_text_range() for page in pdf_pages])
-    elif extension == "docx":
-        doc = docx.Document(BytesIO(file_bytes))
-        return "\n".join([paragraph.text for paragraph in doc.paragraphs])
-    else:
-        text = ""
-        if with_pandoc:
-            try:
-                temporary_job_folder = os.getcwd()
-                with tempfile.NamedTemporaryFile(dir=temporary_job_folder, suffix=".{}".format(extension)) as tmp:
-                    tmp.write(file_bytes)
-                    text = pypandoc.convert_file(tmp.name, to="plain", format=extension)
-
-            except Exception as e:
-                logger.warning("Failed to extract text with pandoc: {}".format(e))
-
-        if not text.strip():
-            return file_bytes.decode()
-        
-        return text
-            
-
