+++ python-lib/text_extraction_ocr_utils/recipes_io_utils.py
